+++ packages/ui/src/lib/data/icons.json
@@ -171,9 +171,9 @@
 	"bowtie-small": "M3 13V3L7.99215 7.37879L13 3V13L7.99215 8.63636L3 13Z",
 	"download": "M4.03039 5.32502L7.75003 9.04466L7.75003 1.85535L9.25003 1.85535L9.25003 9.04472L12.9697 5.32502L14.0304 6.38568L9.73749 10.6786C9.05408 11.362 7.94604 11.362 7.26262 10.6786L2.96973 6.38568L4.03039 5.32502ZM3.50006 14.1053L13.5001 14.1053L13.5001 12.6053L3.50006 12.6053L3.50006 14.1053Z",
 	"arrow-left": "M4.31063 7.23033L8.0303 3.51066L6.96964 2.45L2.67675 6.74289C1.99333 7.42631 1.99333 8.53435 2.67675 9.21776L6.96964 13.5107L8.0303 12.45L4.31063 8.73033L13.5 8.73033L13.5 7.23033L4.31063 7.23033Z",
-	"arrow-top": "M8.76966 4.31063L12.4893 8.0303L13.55 6.96964L9.25709 2.67675C8.57368 1.99333 7.46564 1.99333 6.78222 2.67675L2.48933 6.96964L3.54999 8.0303L7.26966 4.31063L7.26966 13.5L8.76966 13.5L8.76966 4.31063Z",
+	"arrow-up": "M8.76966 4.31063L12.4893 8.0303L13.55 6.96964L9.25709 2.67675C8.57368 1.99333 7.46564 1.99333 6.78222 2.67675L2.48933 6.96964L3.54999 8.0303L7.26966 4.31063L7.26966 13.5L8.76966 13.5L8.76966 4.31063Z",
 	"arrow-right": "M11.6894 8.76967L7.9697 12.4893L9.03036 13.55L13.3233 9.25711C14.0067 8.57369 14.0067 7.46565 13.3233 6.78224L9.03036 2.48934L7.9697 3.55L11.6894 7.26967L2.50003 7.26967L2.50003 8.76967L11.6894 8.76967Z",
-	"arrow-bottom": "M7.23034 11.6894L3.51067 7.9697L2.45001 9.03036L6.74291 13.3233C7.42632 14.0067 8.53436 14.0067 9.21778 13.3233L13.5107 9.03036L12.45 7.9697L8.73034 11.6894L8.73034 2.50003L7.23034 2.50003L7.23034 11.6894Z",
+	"arrow-down": "M7.23034 11.6894L3.51067 7.9697L2.45001 9.03036L6.74291 13.3233C7.42632 14.0067 8.53436 14.0067 9.21778 13.3233L13.5107 9.03036L12.45 7.9697L8.73034 11.6894L8.73034 2.50003L7.23034 2.50003L7.23034 11.6894Z",
 	"fullscreen-resize-enter": "M6 2.25L12 2.25C12.9665 2.25 13.75 3.0335 13.75 4L13.75 10L12.25 10L12.25 4.81066L4.81066 12.25L10 12.25V13.75L4 13.75C3.0335 13.75 2.25 12.9665 2.25 12L2.25 6H3.75L3.75 11.1893L11.1893 3.75L6 3.75V2.25Z",
 	"fullscreen-resize-exit": "M10 7.75L15 7.75L15 6.25L10.8107 6.25L14.6732 2.38747L13.6125 1.32681L9.75 5.18934V1L8.25 1V6C8.25 6.9665 9.0335 7.75 10 7.75ZM6.25 10.8106L6.25 15L7.75 15L7.75 10C7.75 9.0335 6.9665 8.25 6 8.25L1 8.25L1 9.75L5.18933 9.75L1.3268 13.6125L2.38746 14.6732L6.25 10.8106Z",
 	"interdiff": "M6.75 1.5L0 1.5L0 0L6.75 0C7.99264 0 9 1.00736 9 2.25V4.01373C9.08208 4.00466 9.1655 4 9.25 4L16 4V5.5L9.25 5.5C8.83579 5.5 8.5 5.83579 8.5 6.25L8.5 13.75C8.5 14.1642 8.83579 14.5 9.25 14.5H16L16 16H9.25C8.00736 16 7 14.9926 7 13.75V11.9862C6.91784 11.9953 6.83442 12 6.75 12L0 12L0 10.5L6.75 10.5C6.83836 10.5 6.92228 10.485 7 10.4575L7 6.25C7 5.71418 7.1873 5.2221 7.5 4.83567V2.25C7.5 1.83579 7.16421 1.5 6.75 1.5ZM3.5 3L3.5 5.25L-1 5.25L-1 6.75L3.5 6.75V9L6.5 6L3.5 3ZM12.5 7L12.5 9.25H17V10.75H12.5V13L9.5 10L12.5 7Z",
